refactor(db): #566 P2 Stage 2 commit 2 — extract connection.py (sole sqlite3 importer)
Per codex consult layout. connection.py now owns:
  - sqlite3 import (sole — hook regex updated in commit 1)
  - DB_PATH (with corrected .parent depth for new package location)
  - get_connection / get_db / init_db / _apply_migrations

__init__.py shrinks: re-exports from .connection + keeps all other 45
functions (writers, queries, lookups) — those move in commit 3.

### Two test-monkeypatch compat shims (critical)

Found via test failures (99 → 1 → 0):

1. **DB_PATH lookup** in `_resolve_db_path()`: tests do
   `monkeypatch.setattr(db_mod, "DB_PATH", tmp_path)`. Pre-split this worked
   because get_connection read DB_PATH from same module scope. Post-split
   the facade re-export is a separate binding from connection.py's local
   DB_PATH. Helper resolves via facade attribute access (deferred lookup
   ~ negligible cost vs preserving 100+ test fixtures).

2. **get_connection lookup** in `get_db()`: tests/conftest.py autouse
   `_force_no_wal` fixture patches `db_mod.get_connection` to a
   MEMORY-journal version (CI tmpfs cross-connection visibility). Same
   binding issue — get_db() now resolves via facade so the patch works.

Without these shims: 99 test failures (DB_PATH leak to production DB,
journal_mode mismatch causing INSERT invisibility on tmpfs).

Verify:
- import time 208ms (under codex 230ms threshold)
- 4465/4465 tests pass
- make lint clean, make typecheck 0 errors

Next commit: extract behavior modules (market_data/portfolio/audit/
agent_runtime/research_ops/execution_ops/decisions/trades).
